requirements.txt
requests==2.31
python-dotenv==1.0
pyscreenshot==3.1
